requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
h11==0.12.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.22.1
orjson==3.4.7
passlib==1.7.4
promise==2.3
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pymongo==3.11.2
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
Rx==1.6.1
six==1.15.0
starlette==0.13.6
ujson==3.2.0
urllib3==1.26.3
uvicorn==0.13.3
watchgod==0.6
websockets==8.1
